3. web frontend/nda_app/requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
gunicorn==21.2.0

transformers==4.35.2
sentence-transformers==2.2.2
tokenizers==0.15.0
torch==2.1.0
numpy==1.26.2
pandas==2.1.4

xgboost==2.0.3
scikit-learn==1.3.2
joblib==1.3.2

python-docx==1.1.0
requests==2.31.0
